backend/backend: refactor backend accounts keystores
